+++ src/main/java/com/atlauncher/constants/UIConstants.java
@@ -40,4 +40,13 @@ public class UIConstants {
     // When using FlowLayout with a horizonal margin, we need to negate the first
     // components margin added from the FlowLayout
     public static final Insets FLOW_FIELD_INSETS = new Insets(SPACING_LARGE, (-SPACING_LARGE) - 3, SPACING_LARGE, 0);
+
+    public static final int LAUNCHER_NEWS_TAB = 0;
+    public static final int LAUNCHER_VANILLA_PACKS_TAB = 1;
+    public static final int LAUNCHER_PACKS_TAB = 2;
+    public static final int LAUNCHER_INSTANCES_TAB = 3;
+    public static final int LAUNCHER_SERVERS_TAB = 4;
+    public static final int LAUNCHER_ACCOUNTS_TAB = 5;
+    public static final int LAUNCHER_TOOLS_TAB = 6;
+    public static final int LAUNCHER_SETTINGS_TAB = 7;
 }