+++ src/main/java/com/atlauncher/gui/dialogs/ImportInstanceDialog.java
@@ -205,6 +205,10 @@ public class ImportInstanceDialog extends JDialog {
                 DialogManager.okDialog().setTitle(GetText.tr("No Account Selected"))
                         .setContent(GetText.tr("Cannot create instance as you have no account selected."))
                         .setType(DialogManager.ERROR).show();
+
+                if (AccountManager.getAccounts().size() == 0) {
+                    App.launcherFrame.tabbedPane.setSelectedIndex(UIConstants.LAUNCHER_ACCOUNTS_TAB);
+                }
                 return;
             }